+++ indispensable-josef/src/commonMain/kotlin/at/asitplus/signum/indispensable/josef/JweHeader.kt
@@ -3,6 +3,7 @@ package at.asitplus.signum.indispensable.josef
 import at.asitplus.catching
 import at.asitplus.signum.indispensable.io.ByteArrayBase64UrlSerializer
 import at.asitplus.signum.indispensable.io.CertificateChainBase64Serializer
+import at.asitplus.signum.indispensable.josef.algorithm.JweAlgorithm
 import at.asitplus.signum.indispensable.josef.io.InstantLongSerializer
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import at.asitplus.signum.indispensable.pki.CertificateChain
