+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/PrivateKey.kt
@@ -1,0 +1,13 @@
+package at.asitplus.signum.indispensable
+
+@Deprecated(
+    "Moved to at.asitplus.signum.indispensable.key.",
+    ReplaceWith("PrivateKey", "at.asitplus.signum.indispensable.key.PrivateKey")
+)
+typealias PrivateKey = at.asitplus.signum.indispensable.key.PrivateKey
+
+@Deprecated(
+    "Renamed to PrivateKey.",
+    ReplaceWith("PrivateKey", "at.asitplus.signum.indispensable.key.PrivateKey")
+)
+typealias CryptoPrivateKey = at.asitplus.signum.indispensable.key.PrivateKey
