+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/PublicKey.kt
@@ -1,0 +1,25 @@
+package at.asitplus.signum.indispensable
+
+@Deprecated(
+    "Moved to at.asitplus.signum.indispensable.key.",
+    ReplaceWith("PublicKey", "at.asitplus.signum.indispensable.key.PublicKey")
+)
+typealias PublicKey = at.asitplus.signum.indispensable.key.PublicKey
+
+@Deprecated(
+    "Renamed to PublicKey.",
+    ReplaceWith("PublicKey", "at.asitplus.signum.indispensable.key.PublicKey")
+)
+typealias CryptoPublicKey = at.asitplus.signum.indispensable.key.PublicKey
+
+@Deprecated(
+    "Moved to at.asitplus.signum.indispensable.key.",
+    ReplaceWith("SpecializedPublicKey", "at.asitplus.signum.indispensable.key.SpecializedPublicKey")
+)
+typealias SpecializedPublicKey = at.asitplus.signum.indispensable.key.SpecializedPublicKey
+
+@Deprecated(
+    "Renamed to SpecializedPublicKey.",
+    ReplaceWith("SpecializedPublicKey", "at.asitplus.signum.indispensable.key.SpecializedPublicKey")
+)
+typealias SpecializedCryptoPublicKey = at.asitplus.signum.indispensable.key.SpecializedPublicKey
