+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/Signature.kt
@@ -1,0 +1,13 @@
+package at.asitplus.signum.indispensable
+
+@Deprecated(
+    "Moved to at.asitplus.signum.indispensable.signature.",
+    ReplaceWith("Signature", "at.asitplus.signum.indispensable.signature.Signature")
+)
+typealias Signature = at.asitplus.signum.indispensable.signature.Signature
+
+@Deprecated(
+    "Renamed to Signature.",
+    ReplaceWith("Signature", "at.asitplus.signum.indispensable.signature.Signature")
+)
+typealias CryptoSignature = at.asitplus.signum.indispensable.signature.Signature
