+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/key/PublicKey.kt
@@ -1,4 +1,4 @@
-package at.asitplus.signum.indispensable
+package at.asitplus.signum.indispensable.key
 
 import at.asitplus.KmmResult
 import at.asitplus.awesn1.*
@@ -12,6 +12,11 @@ import at.asitplus.signum.indispensable.asn1.DEFAULT_PEM_DECODER
 import at.asitplus.signum.indispensable.asn1.LabelPemDecodable
 import at.asitplus.signum.indispensable.asn1.toAsn1Integer
 import at.asitplus.signum.indispensable.asn1.toBigInteger
+import at.asitplus.signum.indispensable.Awesn1Backed
+import at.asitplus.signum.indispensable.KeyAgreementPublicValue
+import at.asitplus.signum.indispensable.ec.ECCurve
+import at.asitplus.signum.indispensable.ec.ECPoint
+import at.asitplus.signum.indispensable.ec.fromIosEncodedPublicKeyLength
 import at.asitplus.signum.indispensable.misc.ANSIECPrefix
 import at.asitplus.signum.indispensable.misc.ANSIECPrefix.Companion.hasPrefix
 import at.asitplus.signum.internals.checkedAsFn
