+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/pki/X509Certificate.kt
@@ -12,16 +12,16 @@ import at.asitplus.catching
 import at.asitplus.catchingUnwrapped
 import at.asitplus.signum.indispensable.Awesn1Backed
 import at.asitplus.signum.indispensable.EcdsaSignatureMappingFamily
-import at.asitplus.signum.indispensable.PublicKey
-import at.asitplus.signum.indispensable.Signature
 import at.asitplus.signum.indispensable.SignatureAlgorithm
 import at.asitplus.signum.indispensable.X509SignatureAlgorithm
 import at.asitplus.signum.indispensable.asn1.LabelPemDecodable
 import at.asitplus.signum.indispensable.io.Base64Strict
 import at.asitplus.signum.indispensable.io.TransformingSerializerTemplate
+import at.asitplus.signum.indispensable.key.PublicKey
 import at.asitplus.signum.indispensable.requireSignatureAlgorithm
 import at.asitplus.signum.indispensable.requireSupported
 import at.asitplus.signum.indispensable.signatureMappingKeyOrNull
+import at.asitplus.signum.indispensable.signature.Signature
 import at.asitplus.signum.indispensable.toSignatureAlgorithmIdentifier
 import io.matthewnelson.encoding.base64.Base64
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
